refactor(projects): split create project drawer into focused hooks
